Backend/requirements.txt
fastapi
python-multipart
uvicorn
PyAudio
pygame
groq
python-dotenv
aiofiles